refactor: move view to 'views' directory
