+++ myapp/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path
 
-from . import views
+from .views import index_view
 
 urlpatterns = [
-    path('', views.index, name='index'),
+    path('', index_view.index, name='index'),
 ]
