+++ myapp/views/index_view.py
